+++ src/desktopfallback.h
@@ -8,10 +8,12 @@
 #include <QUrl>
 #include <QVariantList>
 
+#include <functional>
 #include <memory>
 
 class QMenu;
 class QWidget;
+class QAction;
 
 enum class DesktopCreateKind {
     Folder,
@@ -30,6 +32,8 @@ public:
     virtual bool programAvailable(const QString &program) const = 0;
     virtual void openUrl(const QUrl &url) = 0;
     virtual void startProgram(const QString &program, const QStringList &arguments) = 0;
+    virtual bool restartShellAvailable() const = 0;
+    virtual void restartShell() = 0;
     virtual void callSessionBus(const QString &service,
                                 const QString &path,
                                 const QString &interface,
@@ -58,9 +62,18 @@ public:
 
     QMenu *menu() const;
 
+    void setActiveWindowActions(bool canClose,
+                                std::function<void()> closeWindow,
+                                bool canForceQuit,
+                                std::function<void()> forceQuitWindow);
+
 private:
     std::unique_ptr<DesktopFallbackActionRunner> m_ownedRunner;
     DesktopFallbackActionRunner *m_runner = nullptr;
     QString m_desktopDirectory;
     std::unique_ptr<QMenu> m_menu;
+    QAction *m_closeWindowAction = nullptr;
+    QAction *m_forceQuitWindowAction = nullptr;
+    std::function<void()> m_closeWindow;
+    std::function<void()> m_forceQuitWindow;
 };